control.lua rework, gamepad stuff, AI stuff
Input now occurs on a frame-by-frame basis, which should make input
streaming possible when we get to networking. Lube is imported now,
you'll notice, but I don't even have lobbies yet. Gamepads used to
crash the game apparently, that has a TODO now. The build script allows
for renames, which is more for the side projects but worth noting.